sem/na: add new symbolDeclare method
I'm really proud of this one.

This magic takes away the pains of casting and performs symbol
checks and instance checks for you.

All you need to do is something like this:

```java
VarSymbol s = symbolRequest(v.name, VarSymbol.class);
```

The class provided to symbolRequest must be a class that extends
Symbol, so it's stupid friendly. Meaning you can't screw up as a
developer and expect the symbol name to refer to a class that
isn't actually a Symbol (which is impossible).

All this is to clean up detection of this malformed C code:

```c
int a;

void main() {
	a();
}
```

Refer to the next commit to see this in action.