gen: fix visitVarExpr for pointers (see desc)
visitVarExpr works by first getting the address of the underlying value.

The value of visitVarExpr are both addresses when the VarExpr refers
to a PointerType or ArrayType. This address should be in a register.

For PointerType, this address is stored as a word someplace in memory.
We need to load this word from memory and set `value` to this word.

For ArrayType, however, this is not the case. ArrayTypes are stored
entirely in memory. The value should be the memory address we load the
word above from -- except we don't load the word! We skip that step.

It's like this:

	somePointer: 0x00000001
	someArray:   1
	             2
		     3
		     4
		     5

The pointer value is an address stored in memory at addressOf(somePointer).

The value of someArray is the memory location of the first item.

No loading!